# openapi.yaml
components:
  schemas:
    ArtworkInput:
      properties:
        description:
          title: Description
          type: string
        id:
          anyOf:
          - type: string
          - type: 'null'
          title: Id
        title:
          title: Title
          type: string
      required:
      - title
      - description
      title: ArtworkInput
      type: object
    ArtworkOutput:
      properties:
        description:
          title: Description
          type: string
        id:
          title: Id
          type: string
        title:
          title: Title
          type: string
      required:
      - id
      - title
      - description
      title: ArtworkOutput
      type: object
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
info:
  description: This is a sample API
  title: Your API
  version: 1.0.0
openapi: 3.0.3
paths:
  /api/v1/artwork/:
    get:
      operationId: list_artwork_api_v1_artwork__get
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/ArtworkOutput'
                title: Response 200 List Artwork Api V1 Artwork  Get
                type: array
          description: Successful Response
      summary: List Artwork
    post:
      operationId: create_artwork_api_v1_artwork__post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ArtworkInput'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArtworkOutput'
          description: Created
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Create Artwork
  /api/v1/artwork/health:
    get:
      operationId: health_api_v1_artwork_health_get
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
      summary: Health
  /api/v1/artwork/{id}:
    delete:
      operationId: delete_artwork_api_v1_artwork__id__delete
      parameters:
      - in: path
        name: id
        required: true
        schema:
          title: Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Delete Artwork
    get:
      operationId: get_artwork_api_v1_artwork__id__get
      parameters:
      - in: path
        name: id
        required: true
        schema:
          title: Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArtworkOutput'
          description: Successful Response
        '404':
          description: Not Found
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Artwork
    patch:
      operationId: update_artwork_api_v1_artwork__id__patch
      parameters:
      - in: path
        name: id
        required: true
        schema:
          title: Id
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ArtworkInput'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArtworkOutput'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Update Artwork